# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  app_internal_interfaces_http_dto_request.BlogCategoryCreateRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 255
        type: string
      slug:
        maxLength: 255
        type: string
    required:
    - name
    - slug
    type: object
  app_internal_interfaces_http_dto_request.BlogCategoryUpdateRequest:
    properties:
      description:
        type: string
      name:
        maxLength: 255
        type: string
      slug:
        maxLength: 255
        type: string
    required:
    - name
    - slug
    type: object
...
  app_internal_interfaces_http_dto_response.AdminBlogCommentListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/app_internal_interfaces_http_dto_response.BlogCommentListItemResponse'
        type: array
      meta:
        $ref: '#/definitions/app_pkg_pagination.Meta'
    type: object
...
  app_internal_interfaces_http_dto_response.BlogCategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      slug:
        type: string
    type: object
...
  app_internal_interfaces_http_dto_response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorBody'
      path:
        example: /api/v1/auth/login
        type: string
      statusCode:
        example: 400
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@shop.com
    name: API Support
  description: REST API for the eCommerce admin panel and public storefront. Supports
    JWT authentication with role-based access (admin, customer). Store endpoints under
    /api/v1/store are public; account routes require a customer token.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: eCommerce API
  version: "1.0"
paths:
...
  /api/v1/admin/blog/categories:
    get:
      description: Get list of all blog categories for admin panel.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/app_internal_interfaces_http_dto_response.BlogCategoryResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List categories (Admin)
      tags:
      - admin-blog
    post:
      consumes:
      - application/json
      description: Create a new blog category.
      parameters:
      - description: Category data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_internal_interfaces_http_dto_request.BlogCategoryCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.BlogCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create category (Admin)
      tags:
      - admin-blog
  /api/v1/admin/blog/categories/{id}:
    delete:
      description: Delete a blog category by ID.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete category (Admin)
      tags:
      - admin-blog
    get:
      description: Get details of a single blog category by ID.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.BlogCategoryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get category (Admin)
      tags:
      - admin-blog
    put:
      consumes:
      - application/json
      description: Update details of an existing blog category.
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_internal_interfaces_http_dto_request.BlogCategoryUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.BlogCategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update category (Admin)
      tags:
      - admin-blog
  /api/v1/admin/blog/comments:
    get:
      description: Get a paginated list of all blog comments.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Post ID filter
        in: query
        name: post_id
        type: string
      - description: Status filter (pending/approved/rejected)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.AdminBlogCommentListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List comments (Admin)
      tags:
      - admin-blog
  /api/v1/admin/blog/comments/{id}:
    delete:
      description: Delete a blog comment by ID.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete comment (Admin)
      tags:
      - admin-blog
  /api/v1/admin/blog/comments/{id}/status:
    patch:
      consumes:
      - application/json
      description: Approve or reject a blog comment by ID.
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      - description: Moderation data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_internal_interfaces_http_dto_request.BlogCommentModerateRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Moderate comment (Admin)
      tags:
      - admin-blog
  /api/v1/admin/blog/posts:
    get:
      description: Get a paginated list of all posts (drafts and published).
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Search query
        in: query
        name: q
        type: string
      - description: Category ID filter
        in: query
        name: category_id
        type: string
      - description: Status filter (draft/published)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.BlogPostListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List posts (Admin)
      tags:
      - admin-blog
    post:
      consumes:
      - application/json
      description: Create a new blog post.
      parameters:
      - description: Post data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/app_internal_interfaces_http_dto_request.BlogPostCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.BlogPostResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create post (Admin)
      tags:
      - admin-blog
...
  /api/v1/store/blog/posts:
    get:
      description: Get a paginated list of published blog posts.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Search query
        in: query
        name: q
        type: string
      - description: Category ID filter
        in: query
        name: category_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.BlogPostListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      summary: List blog posts
      tags:
      - blog
...
  /api/v1/store/blog/posts/{slug}:
    get:
      description: Get details of a single published blog post by slug.
      parameters:
      - description: Post slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.BlogPostResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app_internal_interfaces_http_dto_response.ErrorResponse'
      summary: Get blog post
      tags:
      - blog
...
securityDefinitions:
  BearerAuth:
    description: 'Enter "Bearer" followed by a space and the JWT access token. Example:
      Bearer eyJhbGciOiJIUzI1NiIs...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"